Make the script wait for events as a parameter.
I still shouldn't be using write-host in the event action but for now it's
at least making this script function.

without `-Wait` we will return to the repl but the event will still be
registerd. Writing to your host when it is fired.

Begin now unregisters and active events when called in order to replace
the event hook. This avoids creating lots of listening events.