Access type lookup table and Stream output into a file
Before was buffering the events into a string and would then write the
string to a file. This uses all the RAM then crashes. This commit now
creates a file and gets a stream object to this file to continuously
append data to.

The accessMask just returns a hard to read HEX number, This adds a small
table to reference this number for some output in english.